P2P: ignore neg_req with previously used dialog_token
The p2p spec is a bit broken. if for some reason we
get duplicate negotiation request, the supplicant
will generate 2 different responses (different SSIDs)
with the same dialog token. The remote peer will
confirm one of them, but it will probably be the
wrong one (the first he received)...

Workaround it but ignoring negotiation requests with
the last used dialog_token.

Signed-off-by: Eliad Peller <eliad@wizery.com>