sched scan: Make initial short interval scans more aggressive (INTERNAL)
In a very congested environment we've encountered an issue
where a remembered AP isn't detected in the first or second
scan necessarily. The current default of 10 seconds as the short
sched scan interval would therefore cause a long delay until
connecting to a remembered AP after powering on WiFi or disconnecting.
To fix this, reduce the default short interval to 5 secs and
have 10 intervals instead of 6 intervals of 10 seconds. The effect
on power consumption shouldn't be significant as the long interval
is the important factor affecting that.

Signed-off-by: Eyal Shapira <eyal@wizery.com>